// backend/src/modules/logistics/enums/delivery-failure-reason.enum.ts
export enum DeliveryFailureReason {
  CUSTOMER_UNAVAILABLE = 'CUSTOMER_UNAVAILABLE',
  INCORRECT_ADDRESS = 'INCORRECT_ADDRESS',
  CUSTOMER_REFUSED_DELIVERY = 'CUSTOMER_REFUSED_DELIVERY',
  CUSTOMER_REFUSED_PAYMENT = 'CUSTOMER_REFUSED_PAYMENT',
  VEHICLE_BREAKDOWN = 'VEHICLE_BREAKDOWN',
  SHOP_CLOSED_OR_UNAVAILABLE = 'SHOP_CLOSED_OR_UNAVAILABLE',
  ITEMS_DAMAGED = 'ITEMS_DAMAGED',
}